% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_paths.R
\name{get_sub_paths}
\alias{get_sub_paths}
\title{Get sub paths}
\usage{
get_sub_paths(path, pattern, col_name = "abs_path")
}
\arguments{
\item{path}{A path.}

\item{pattern}{A regex.}

\item{col_name}{A string.}
}
\value{
tibble::tibble with character column col_name containing abs_paths of subfiles
}
\description{
Get sub paths
}
\examples{
get_sub_paths("./data", pattern = ".csv", col_name = "abs_path")
library(readutils); readutils::get_sub_paths(path = "/Users/sorpet/Desktop/data",
pattern = ".csv",
col_name = "abs_path")
}
